add ability to constrain specific lens model parameters during the optimization; code restructuring and documentation
