+++ lenstronomy/LensModel/Optimizer/multi_plane.py
@@ -1,192 +1,339 @@
-from lenstronomy.LensModel.Solver.lens_equation_solver import LensEquationSolver
-from lenstronomy.LensModel.Optimizer.split_multiplane import SplitMultiplane as split_multi
+from lenstronomy.LensModel.lens_model import LensModel
 import numpy as np
 
+class MultiPlaneLensing(object):
 
-class MultiPlaneOptimizer(object):
+    def __init__(self, full_lensmodel, x_pos, y_pos, lensmodel_params, z_source,
+                 z_macro, astropy_instance, macro_indicies):
 
-    def __init__(self, lensmodel_full, all_args, x_pos, y_pos, tol_source, Params, magnification_target,
-                 tol_mag, centroid_0, tol_centroid, z_main, z_src, astropy_instance, interpolated,
-                 mag_penalty=False,verbose=False,pso_convergence_mean=None,pso_compute_magnification=None):
+        """
+        This class performs (fast) lensing computations for multi-plane lensing scenarios
+        :param full_lensmodel:
+        :param x_pos:
+        :param y_pos:
+        :param lensmodel_params:
+        :param z_source:
+        :param z_macro:
+        :param astropy_instance:
+        :param macro_indicies:
+        """
 
-        self.Params = Params
-        self.lensModel = lensmodel_full
-        self.all_lensmodel_args = all_args
-        self.solver = LensEquationSolver(self.lensModel)
+        self.z_macro, self.z_source = z_macro, z_source
 
-        self.tol_source = tol_source
+        self.astropy_instance = astropy_instance
 
-        self.magnification_target = magnification_target
-        self.tol_mag = tol_mag
+        self.x_pos, self.y_pos = np.array(x_pos), np.array(y_pos)
 
-        self._compute_mags_flag = mag_penalty
+        self.full_lensmodel, self.lensmodel_params = full_lensmodel, lensmodel_params
 
-        self._pso_convergence_mean = pso_convergence_mean
-        self._pso_compute_magnification = pso_compute_magnification
+        self._T_z_source = full_lensmodel.lens_model._T_z_source
 
-        self.centroid_0 = centroid_0
-        self.tol_centroid = tol_centroid
+        macromodel_lensmodel, macro_args, halo_lensmodel, halo_args, self._z_background = \
+            self._split_lensmodel(full_lensmodel,lensmodel_params,z_break=z_macro,macro_indicies=macro_indicies)
 
-        self.verbose = verbose
+        self.foreground = Foreground(halo_lensmodel,self.z_macro,x_pos,y_pos)
+        self.halo_args = halo_args
 
-        self._x_pos,self._y_pos = np.array(x_pos),np.array(y_pos)
-        self.mag_penalty,self.src_penalty,self.parameters = [],[], []
+        self.model_to_vary = ToVary(macromodel_lensmodel,self.z_macro,self._z_background)
+        self.macro_args = macro_args
 
-        self.multiplane_optimizer = split_multi(x_pos,y_pos,lensmodel_full,all_args,interpolated=interpolated,
-                                                    z_source=z_src,z_macro=z_main,astropy_instance=astropy_instance,verbose=verbose,
-                                                    macro_indicies=self.Params.tovary_indicies)
+        self.background = Background(halo_lensmodel,self.z_macro,self.z_source)
 
-        self.reset()
+    def ray_shooting(self, thetax, thetay, macromodel_args):
 
-    def reset(self,compute_mags=False):
+        # get the deflection angles from foreground and main lens plane subhalos (once)
+        x, y, alphax, alphay = self.foreground.ray_shooting(self.halo_args,thetax=thetax,thetay=thetay)
 
-        self.mag_penalty, self.src_penalty, self.centroid_penalty, self.parameters = [], [], [], []
-        self._converged = False
-        self._counter = 1
-        self._compute_mags = compute_mags
-        self.is_converged = False
+        x, y, alphax, alphay = self.model_to_vary.ray_shooting(alphax, alphay, macromodel_args, x, y)
 
-    def get_best(self):
+        # compute the angular position on the source plane
+        x_source, y_source, _, _ = self.background.ray_shooting(alphax, alphay, self.halo_args, x, y)
 
-        total = np.array(self.src_penalty)+np.array(self.mag_penalty)+np.array(self.centroid_penalty)
-        index = np.argmin(total)
+        betax, betay = x_source * self._T_z_source ** -1, y_source * self._T_z_source ** -1
 
-        return total[index]
+        return betax, betay
 
-    def _init_particles(self,n_particles,n_iterations):
+    def ray_shooting_fast(self, macromodel_args, true_foreground=True, offset_index=None,thetax=None,thetay=None):
 
-        self._n_total_iter = n_iterations*n_particles
-        self._n_particles = n_particles
-        self._mag_penalty_switch = 1
+        # get the deflection angles from foreground and main lens plane subhalos (once)
+        x, y, alphax, alphay = self.foreground.ray_shooting(self.halo_args,true_foreground=true_foreground,
+                                                            offset_index=offset_index,thetax=thetax,thetay=thetay
+                                                            ,force_compute=False)
 
-    def _get_images(self,kwargs_varied):
+        x, y, alphax, alphay = self.model_to_vary.ray_shooting(alphax, alphay, macromodel_args, x, y)
 
-        srcx, srcy = self.multiplane_optimizer.ray_shooting_fast(kwargs_varied)
+        # compute the angular position on the source plane
+        x_source, y_source, _, _ = self.background.ray_shooting(alphax, alphay, self.halo_args, x, y)
 
-        args = kwargs_varied + self.Params.argsfixed_todictionary()
+        betax, betay = x_source * self._T_z_source ** -1, y_source * self._T_z_source ** -1
 
-        source_x, source_y = np.mean(srcx), np.mean(srcy)
-        x_image, y_image = self.solver.image_position_from_source(source_x, source_y, args, precision_limit=10**-10)
+        return betax, betay
 
-        return x_image, y_image, source_x, source_y
+    def magnification(self,thetax,thetay,macromodel_args):
 
-    def _source_position_penalty(self, lens_args_tovary):
+        fxx,fxy,fyx,fyy = self.hessian(thetax,thetay,macromodel_args)
 
-        betax,betay = self.multiplane_optimizer.ray_shooting_fast(lens_args_tovary)
+        det_J = (1-fxx)*(1-fyy)-fyx*fxy
 
-        dx = ((betax[0] - betax[1]) ** 2 + (betax[0] - betax[2]) ** 2 + (betax[0] - betax[3]) ** 2 + (
-                betax[1] - betax[2]) ** 2 +
-              (betax[1] - betax[3]) ** 2 + (betax[2] - betax[3]) ** 2)
-        dy = ((betay[0] - betay[1]) ** 2 + (betay[0] - betay[2]) ** 2 + (betay[0] - betay[3]) ** 2 + (
-                betay[1] - betay[2]) ** 2 +
-              (betay[1] - betay[3]) ** 2 + (betay[2] - betay[3]) ** 2)
-        return 0.5 * (dx + dy) * self.tol_source ** -2
+        return np.absolute(det_J**-1)
 
-    def _magnification_penalty(self, lens_args,magnification_target, tol):
+    def magnification_fast(self,macromodel_args):
 
-        magnifications = self.multiplane_optimizer.magnification_fast(lens_args)
+        fxx,fxy,fyx,fyy = self.hessian_fast(macromodel_args)
 
-        magnifications *= max(magnifications) ** -1
+        det_J = (1-fxx)*(1-fyy)-fyx*fxy
 
-        dM = []
+        return np.absolute(det_J**-1)
 
-        for i, target in enumerate(magnification_target):
-            mag_tol = tol * target
-            dM.append((magnifications[i] - target) * mag_tol ** -1)
+    def hessian(self,thetax,thetay,macromodel_args,diff=0.00000001):
 
-        dM = np.array(dM)
+        alpha_ra, alpha_dec = self._alpha(thetax,thetay, macromodel_args)
 
-        return 0.5 * np.sum(dM ** 2)
+        alpha_ra_dx, alpha_dec_dx = self._alpha(thetax + diff, thetay, macromodel_args)
+        alpha_ra_dy, alpha_dec_dy = self._alpha(thetax, thetay + diff, macromodel_args)
 
-    def _centroid_penalty(self, values_dic, tol_centroid):
+        dalpha_rara = (alpha_ra_dx - alpha_ra) * diff ** -1
+        dalpha_radec = (alpha_ra_dy - alpha_ra) * diff ** -1
+        dalpha_decra = (alpha_dec_dx - alpha_dec) * diff ** -1
+        dalpha_decdec = (alpha_dec_dy - alpha_dec) * diff ** -1
 
-        d_centroid = ((values_dic[0]['center_x'] - self.centroid_0[0]) * tol_centroid ** -1) ** 2 + \
-                     ((values_dic[0]['center_y'] - self.centroid_0[1]) * tol_centroid ** -1) ** 2
+        f_xx = dalpha_rara
+        f_yy = dalpha_decdec
+        f_xy = dalpha_radec
+        f_yx = dalpha_decra
 
-        return 0.5 * d_centroid
+        return f_xx, f_xy, f_yx, f_yy
 
-    def _log(self,src_penalty,mag_penalty,centroid_penalty):
+    def hessian_fast(self,macromodel_args,diff=0.00000001):
 
-        if mag_penalty is None:
-            mag_penalty = 10**10
-        if src_penalty is None:
-            src_penalty = 10**10
-        if centroid_penalty is None:
-            centroid_penalty = 10**10
+        alpha_ra, alpha_dec = self._alpha_fast(self.x_pos,self.y_pos, macromodel_args, true_foreground=True)
 
-        self.src_penalty.append(np.sum(src_penalty))
-        self.mag_penalty.append(np.sum(mag_penalty))
-        self.centroid_penalty.append(np.sum(centroid_penalty))
+        alpha_ra_dx, alpha_dec_dx = self._alpha_fast(self.x_pos + diff, self.y_pos, macromodel_args, true_foreground=False,
+                                                     offset_index=0)
+        alpha_ra_dy, alpha_dec_dy = self._alpha_fast(self.x_pos, self.y_pos + diff, macromodel_args, true_foreground=False,
+                                                     offset_index=1)
 
-        self.parameters.append(self.lens_args_latest)
+        dalpha_rara = (alpha_ra_dx - alpha_ra) * diff ** -1
+        dalpha_radec = (alpha_ra_dy - alpha_ra) * diff ** -1
+        dalpha_decra = (alpha_dec_dx - alpha_dec) * diff ** -1
+        dalpha_decdec = (alpha_dec_dy - alpha_dec) * diff ** -1
 
-    def _compute_mags_criterion(self):
+        f_xx = dalpha_rara
+        f_yy = dalpha_decdec
+        f_xy = dalpha_radec
+        f_yx = dalpha_decra
 
-        if self._compute_mags:
-            return True
+        return f_xx, f_xy, f_yx, f_yy
 
-        if self._counter <= self._n_particles:
-            return False
+    def _alpha(self, x_pos, y_pos, macromodel_args):
 
-        if np.mean(self.src_penalty[-self._n_particles:]) < self._pso_compute_magnification:
-            return True
+        beta_x,beta_y = self.ray_shooting(x_pos,y_pos,macromodel_args)
+
+        alpha_x = np.array(x_pos - beta_x)
+        alpha_y = np.array(y_pos - beta_y)
+
+        return alpha_x, alpha_y
+
+    def _alpha_fast(self, x_pos, y_pos, macromodel_args, true_foreground=False, offset_index = None):
+
+        beta_x,beta_y = self.ray_shooting_fast(macromodel_args, true_foreground=true_foreground, offset_index=offset_index,
+                                               thetax=x_pos,thetay=y_pos)
+
+        alpha_x = np.array(x_pos - beta_x)
+        alpha_y = np.array(y_pos - beta_y)
+
+        return alpha_x, alpha_y
+
+    def _split_lensmodel(self, lensmodel, lensmodel_args, z_break, macro_indicies):
+
+        """
+
+        :param lensmodel: lensmodel to break up
+        :param lensmodel_args: kwargs to break up
+        :param z_break: the break redshift
+        :param macro_indicies: the indicies of the macromodel in the lens model list
+        :return: instances of LensModel for foreground, main lens plane and background halos, and the macromodel
+        """
+
+        front_model_names, front_redshifts, front_args = [], [], []
+        back_model_names, back_redshifts, back_args = [], [], []
+        macro_names, macro_redshifts, macro_args = [], [], []
+
+        halo_names, halo_redshifts, halo_args = [], [], []
+
+        background_z_current = self.z_macro + 0.5*(self.z_source - self.z_macro)
+
+        for i in range(0, len(lensmodel.lens_model_list)):
+
+            z = lensmodel.redshift_list[i]
+
+            if i not in macro_indicies:
+
+                halo_names.append(lensmodel.lens_model_list[i])
+                halo_redshifts.append(z)
+                halo_args.append(lensmodel_args[i])
+
+                if z > z_break:
+
+                    if z < background_z_current:
+                        background_z_current = z
+
+                    back_model_names.append(lensmodel.lens_model_list[i])
+                    back_redshifts.append(z)
+                    back_args.append(lensmodel_args[i])
+
+                elif z <= z_break:
+                    front_model_names.append(lensmodel.lens_model_list[i])
+                    front_redshifts.append(z)
+                    front_args.append(lensmodel_args[i])
+
+            else:
+
+                macro_names.append(lensmodel.lens_model_list[i])
+                macro_redshifts.append(z)
+                macro_args.append(lensmodel_args[i])
+
+        macromodel = LensModel(lens_model_list=macro_names, redshift_list=macro_redshifts, cosmo=self.astropy_instance,
+                               multi_plane=True,
+                               z_source=self.z_source)
+
+        halo_lensmodel = LensModel(lens_model_list=front_model_names+back_model_names, redshift_list=front_redshifts+back_redshifts,
+                                   cosmo=self.astropy_instance,multi_plane=True,z_source=self.z_source)
+        halo_args = front_args+back_args
+
+        return macromodel, macro_args, halo_lensmodel, halo_args, background_z_current
+
+class ToVary(object):
+
+    def __init__(self,tovary_lensmodel,z_to_vary,z_next_plane):
+
+        self.tovary_lensmodel = tovary_lensmodel
+        self.z_to_vary = z_to_vary
+        #self.z_background = z_next_plane
+
+    def ray_shooting(self, thetax, thetay, args, x_in, y_in):
+
+        x, y, alphax, alphay = self.tovary_lensmodel.lens_model. \
+            ray_shooting_partial(x_in, y_in, thetax, thetay, z_start=self.z_to_vary,
+                                 z_stop=self.z_to_vary, kwargs_lens=args, include_z_start=True)
+
+        return x, y, alphax, alphay
+
+class Foreground(object):
+
+    def __init__(self, foreground_lensmodel, z_to_vary, x_pos, y_pos):
+
+        self.halos_lensmodel = foreground_lensmodel
+        self.z_to_vary = z_to_vary
+        self.x_pos,self.y_pos = x_pos,y_pos
+        self.diff_rays = [None]*2
+
+    def set_precomputed(self,true_rays,offset_rays):
+
+        if true_rays is not None:
+            self.rays = true_rays
+        if offset_rays is not None:
+            self.diff_rays = offset_rays
+
+    def ray_shooting(self,args,true_foreground=False,offset_index=None,thetax=None,thetay=None,force_compute=True):
+
+        if true_foreground:
+
+            if not hasattr(self,'rays'):
+                x0, y0 = np.zeros_like(self.x_pos), np.zeros_like(self.y_pos)
+                x,y,alphax,alphay = self.halos_lensmodel.lens_model.ray_shooting_partial(x0, y0, self.x_pos, self.y_pos,
+                                                                                                 z_start=0,
+                                                                                                 z_stop=self.z_to_vary,
+                                                                                                 kwargs_lens=args)
+                self.rays = {'x':x,'y':y,'alphax':alphax,'alphay':alphay}
+
+            return self.rays['x'],self.rays['y'],self.rays['alphax'],self.rays['alphay']
+
+        elif force_compute:
+
+            x0, y0 = np.zeros_like(self.x_pos), np.zeros_like(self.y_pos)
+            x,y,alphax,alphay = self.halos_lensmodel.lens_model.ray_shooting_partial(x0, y0, thetax, thetay,
+                                                         z_start=0,z_stop=self.z_to_vary,kwargs_lens=args)
+            return x,y,alphax,alphay
+
         else:
-            return False
 
-    def _test_convergence(self):
+            if self.diff_rays[offset_index] is None:
 
-        if self._counter <= self._n_particles:
-            self.is_converged = False
-            return
+                x0, y0 = np.zeros_like(self.x_pos), np.zeros_like(self.y_pos)
+                x, y, alphax, alphay = self.halos_lensmodel.lens_model.ray_shooting_partial(x0, y0, thetax,
+                                                     thetay,z_start=0,z_stop=self.z_to_vary,kwargs_lens=args)
 
-        if np.mean(self.src_penalty[-self._n_particles:]) < self._pso_convergence_mean:
-            self.is_converged = True
-        else:
-            self.is_converged = False
+                self.diff_rays[offset_index] = {'x': x, 'y': y, 'alphax': alphax, 'alphay': alphay}
 
-    def __call__(self, lens_values_tovary, src_penalty=None,mag_penalty=None,centroid_penalty=None):
+            return self.diff_rays[offset_index]['x'],self.diff_rays[offset_index]['y'],self.diff_rays[offset_index]['alphax'],\
+                                self.diff_rays[offset_index]['alphay']
 
-        self._counter += 1
+class Background(object):
 
-        params_fixed = self.Params.argsfixed_todictionary()
-        lens_args_tovary = self.Params.argstovary_todictionary(lens_values_tovary)
+    def __init__(self, background_lensmodel, z_background, z_source):
 
-        if self.tol_source is not None:
+        self.halos_lensmodel = background_lensmodel
+        self.z_background = z_background
+        self.z_source = z_source
 
-            src_penalty = self._source_position_penalty(lens_args_tovary)
+    def ray_shooting(self, thetax, thetay, args, x_in, y_in):
 
-            self._compute_mags = self._compute_mags_criterion()
+        x, y, alphax, alphay = self.halos_lensmodel.lens_model.ray_shooting_partial(x_in,
+                                   y_in, thetax, thetay,z_start=self.z_background,z_stop=self.z_source,kwargs_lens=args)
 
-        if self._compute_mags and self.tol_mag is not None:
-            mag_penalty = self._magnification_penalty(lens_args_tovary,self.magnification_target,self.tol_mag)
+        return x,y,alphax,alphay
 
-        if self.tol_centroid is not None:
-            centroid_penalty = self._centroid_penalty(lens_args_tovary,self.tol_centroid)
 
-        _penalty = [src_penalty,mag_penalty,centroid_penalty]
+# some old code for interpolating the background field, might as well save it for now....
+"""
+    def _interpolate(self, background_lensmodel, x, y):
 
-        penalty = 0
-        for pen in _penalty:
-            if pen is not None:
-                penalty += pen
+        raise Exception('not yet implemented')
+        T_z_interp = background_lensmodel.lens_model._T_z_list[0]
+        self.interp_models = []
+        self.interp_args = []
 
-        if self._counter % 1000 == 0 and self.verbose:
+        x_values, y_values = np.linspace(-self._interp_range, self._interp_range, self._interp_steps), \
+                             np.linspace(-self._interp_range, self._interp_range, self._interp_steps)
 
-            print('source penalty: ', src_penalty)
-            print('centroid penalty: ', centroid_penalty)
+        for count, (xi, yi) in enumerate(zip(x, y)):
 
-            if self.mag_penalty is not None:
-                print('mag penalty: ', mag_penalty)
+            if self.verbose:
+                print('interpolating field behind image ' + str(count + 1) + '...')
 
-        self.lens_args_latest = lens_args_tovary + params_fixed
+            interp_model_i, interp_args_i = self._lensmodel_interpolated((x_values + xi) * T_z_interp ** -1,
+                                                                         (y_values + yi) * T_z_interp ** -1,
+                                                                         self.background_lensmodel,
+                                                                         self.background_args)
 
-        self._log(src_penalty,mag_penalty,centroid_penalty)
+            self.interp_models.append(interp_model_i)
+            self.interp_args.append(interp_args_i)
 
-        self._test_convergence()
+        return self.interp_models, self.interp_args
 
-        return penalty
+    def _lensmodel_interpolated(self, x_values, y_values, interp_lensmodel, interp_args):
 
-        #if self._return_mode == 'PSO':
-        #    return -1 * penalty, None
-        #else:
-        #    return np.array(penalty)
+       
+        :param x_values: 1d array of x coordinates to interpolate
+        :param y_values: 1d array of y coordinates to interpolate
+        (e.g. np.linspace(ymin,ymax,steps))
+        :param interp_lensmodel: lensmodel to interpolate
+        :param interp_args: kwargs for interp_lensmodel
+        :return: interpolated lensmodel
+       
+        xx, yy = np.meshgrid(x_values, y_values)
+        L = int(len(x_values))
+        xx, yy = xx.ravel(), yy.ravel()
+
+        f_x, f_y = interp_lensmodel.alpha(xx, yy, interp_args)
+
+        interp_args = [{'f_x': f_x.reshape(L, L), 'f_y': f_y.reshape(L, L),
+                        'grid_interp_x': x_values, 'grid_interp_y': y_values}]
+
+        return LensModel(lens_model_list=['INTERPOL'], redshift_list=[self._z_background], cosmo=self.astropy_instance,
+                         z_source=self.z_source, multi_plane=True), interp_args
+"""
+
+
+
