+++ lenstronomy/LensModel/Optimizer/single_plane.py
@@ -1,37 +1,24 @@
 import numpy as np
-from lenstronomy.LensModel.Solver.lens_equation_solver import LensEquationSolver
 
+class SinglePlaneLensing(object):
 
-class SinglePlaneOptimizer(object):
+    def __init__(self, lensmodel, x_pos, y_pos, params, arg_list):
 
-    def __init__(self, lensmodel, x_pos, y_pos, tol_source, params,
-                 magnification_target, tol_mag, centroid_0, tol_centroid, k_start=0, arg_list=[],
-                 return_mode='PSO',verbose=False,mag_penalty=False,pso_convergence_mean=None,
-                 pso_compute_magnification=None):
+        """
+        This class performs (fast) lensing computations for single plane scenarios
+        :param lensmodel:
+        :param x_pos:
+        :param y_pos:
+        :param params:
+        :param arg_list:
+        """
 
         self.Params = params
         self.lensModel = lensmodel
-        self.solver = LensEquationSolver(self.lensModel)
-
-        self.tol_source = tol_source
-
-        self.magnification_target = magnification_target
-        self.tol_mag = tol_mag
-        self._compute_mags_flag = mag_penalty
-
-        self.centroid_0 = centroid_0
-        self.tol_centroid = tol_centroid
-
-        self._pso_convergence_mean = pso_convergence_mean
-        self._pso_compute_magnification = pso_compute_magnification
-
         self._x_pos, self._y_pos = np.array(x_pos), np.array(y_pos)
 
-        self.verbose=verbose
-
         self.all_lensmodel_args = arg_list
-
-        self._return_mode = return_mode
+        k_start = params.k_start
 
         # compute the foreground deflections and second derivatives from subhalos
         if k_start > 0 and len(arg_list)>k_start:
@@ -42,51 +29,12 @@ class SinglePlaneOptimizer(object):
             # subhalo deflections
             self.alpha_x_sub, self.alpha_y_sub = self.lensModel.alpha(x_pos,y_pos,arg_list,self._k_sub)
 
-            # subhalo hessian components
-            if tol_mag is not None:
-                self.sub_fxx, self.sub_fxy, _, self.sub_fyy = self.lensModel.hessian(x_pos,y_pos,arg_list,self._k_sub)
-
         else:
-
             self._k_macro, self._k_sub = None, None
             self.alpha_x_sub, self.alpha_y_sub = 0, 0
             self.sub_fxx, self.sub_fyy, self.sub_fxy = 0,0,0
 
-        self._counter = 1
-        self.reset()
-
-    def reset(self,compute_mags=False):
-
-        self.mag_penalty, self.src_penalty, self.parameters, self.centroid_penalty = [], [], [], []
-        self._counter = 1
-        self._compute_mags = compute_mags
-        self.is_converged = False
-
-    def get_best(self):
-
-        total = np.array(self.src_penalty) + np.array(self.mag_penalty) + np.array(self.centroid_penalty)
-
-        return total[np.argmin(total)]
-
-    def _init_particles(self,n_particles,n_iterations):
-
-        self._n_total_iter = n_iterations*n_particles
-        self._n_particles = n_particles
-        self._mag_penalty_switch = 1
-
-    def _get_images(self,kwargs_varied):
-
-        args = kwargs_varied + self.Params.argsfixed_todictionary()
-
-        srcx, srcy = self.lensModel.ray_shooting(self._x_pos, self._y_pos, args, None)
-
-        source_x, source_y = np.mean(srcx), np.mean(srcy)
-
-        x_image, y_image = self.solver.findBrightImage(source_x, source_y,args,precision_limit=10**-10)
-
-        return x_image, y_image, source_x, source_y
-
-    def _source_position_penalty(self, lens_args):
+    def ray_shooting_fast(self, lens_args):
 
         # compute the macromodel deflection
         alphax_macro,alphay_macro = self.lensModel.alpha(self._x_pos,self._y_pos,lens_args,k=self._k_macro)
@@ -95,16 +43,13 @@ class SinglePlaneOptimizer(object):
         betax = self._x_pos - alphax_macro - self.alpha_x_sub
         betay = self._y_pos - alphay_macro - self.alpha_y_sub
 
-        dx = ((betax[0] - betax[1]) ** 2 + (betax[0] - betax[2]) ** 2 + (betax[0] - betax[3]) ** 2 + (
-                    betax[1] - betax[2]) ** 2 +
-              (betax[1] - betax[3]) ** 2 + (betax[2] - betax[3]) ** 2)
-        dy = ((betay[0] - betay[1]) ** 2 + (betay[0] - betay[2]) ** 2 + (betay[0] - betay[3]) ** 2 + (
-                    betay[1] - betay[2]) ** 2 +
-              (betay[1] - betay[3]) ** 2 + (betay[2] - betay[3]) ** 2)
+        return betax,betay
 
-        return 0.5*(dx+dy)*self.tol_source**-2
+    def magnification_fast(self,  args):
 
-    def _magnification_penalty(self,  args, magnification_target, tol=0.1):
+        if not hasattr(self,'sub_fxx'):
+            self.sub_fxx, self.sub_fxy, _, self.sub_fyy = self.lensModel.hessian(self._x_pos, self._y_pos,
+                                                                  self.all_lensmodel_args, self._k_sub)
 
         fxx_macro,fxy_macro,_,fyy_macro = self.lensModel.hessian(self._x_pos, self._y_pos, args, k=self._k_macro)
 
@@ -118,102 +63,4 @@ class SinglePlaneOptimizer(object):
 
         magnifications *= max(magnifications)**-1
 
-        dM = []
-
-        for i, target in enumerate(magnification_target):
-            mag_tol = tol * target
-            dM.append((magnifications[i] - target) * mag_tol ** -1)
-
-        dM = np.array(dM)
-
-        return 0.5*np.sum(dM ** 2)
-
-    def _centroid_penalty(self, values_dic, tol_centroid):
-
-        dx = (values_dic[0]['center_x'] - self.centroid_0[0])*tol_centroid**-1
-        dy = (values_dic[0]['center_y'] - self.centroid_0[1])*tol_centroid**-1
-
-        return 0.5*(dx**2+dy**2)
-
-    def _log(self,src_penalty,mag_penalty,centroid_penalty):
-
-        if mag_penalty is None:
-            mag_penalty = 10**10
-        if src_penalty is None:
-            src_penalty = 10**10
-        if centroid_penalty is None:
-            centroid_penalty = 10**10
-
-        self.src_penalty.append(np.sum(src_penalty))
-        self.mag_penalty.append(np.sum(mag_penalty))
-        self.centroid_penalty.append(np.sum(centroid_penalty))
-
-        self.parameters.append(self.lens_args_latest)
-
-    def _compute_mags_criterion(self):
-
-        if self._compute_mags:
-            return True
-
-        if self._counter <= self._n_particles:
-            return False
-
-        if np.mean(self.src_penalty[-self._n_particles:]) < self._pso_compute_magnification:
-            return True
-        else:
-            return False
-
-    def _test_convergence(self):
-
-        if self._counter <= self._n_particles:
-            self.is_converged = False
-            return
-
-        if np.mean(self.src_penalty[-self._n_particles:]) < self._pso_convergence_mean:
-            self.is_converged = True
-        else:
-            self.is_converged = False
-
-    def __call__(self, lens_values_tovary,src_penalty=None,mag_penalty=None,centroid_penalty=None):
-
-        self._counter += 1
-
-        params_fixed = self.Params.argsfixed_todictionary()
-        lens_args_tovary = self.Params.argstovary_todictionary(lens_values_tovary)
-
-        if self.tol_source is not None:
-            src_penalty = self._source_position_penalty(lens_args_tovary+params_fixed)
-
-            self._compute_mags = self._compute_mags_criterion()
-
-        if self._compute_mags and self.tol_mag is not None:
-            mag_penalty = self._magnification_penalty(lens_args_tovary + params_fixed, self.magnification_target,
-                                                      self.tol_mag)
-
-        if self.tol_centroid is not None:
-            centroid_penalty = self._centroid_penalty(lens_args_tovary, self.tol_centroid)
-
-        _penalty = [src_penalty, mag_penalty, centroid_penalty]
-
-        penalty = 0
-        for pen in _penalty:
-            if pen is not None:
-                penalty += pen
-
-        if self._counter % 500 == 0 and self.verbose:
-
-            print('source penalty: ', src_penalty)
-            print('centroid penalty: ', centroid_penalty)
-
-            if self.mag_penalty is not None:
-                print('mag penalty: ', mag_penalty)
-
-        self.lens_args_latest = lens_args_tovary + params_fixed
-        self._log(src_penalty, mag_penalty, centroid_penalty)
-        self._test_convergence()
-
-        return penalty
-        #if self._return_mode == 'PSO':
-        #    return -1 * penalty, None
-        #else:
-        #    return penalty
+        return magnifications
